+++ src/Narvalo.Brouillons/Edu/Monads/Samples/MonadOr.g.cs
@@ -617,7 +617,7 @@ namespace Narvalo.Edu.Monads.Samples
             return @this.Coalesce(predicate, MonadOr<TResult>.None, other);
         }
 
-        public static void Apply<TSource>(
+        public static void Invoke<TSource>(
             this MonadOr<TSource> @this,
             Action<TSource> action)
         {
@@ -635,10 +635,10 @@ namespace Narvalo.Edu.Monads.Samples
             Require.Object(@this);
             Require.NotNull(action, "action");
 
-            @this.Then(MonadOr.Unit).Apply(_ => action.Invoke());
+            @this.Then(MonadOr.Unit).Invoke(_ => action.Invoke());
         }
 
-        public static void Apply<TSource>(
+        public static void Invoke<TSource>(
             this MonadOr<TSource> @this,
             Action<TSource> action,
             Action caseNone)
@@ -1053,7 +1053,7 @@ namespace Narvalo.Edu.Monads.Samples.Internal
 
                 if (m != null)
                 {
-                    m.Apply(
+                    m.Invoke(
                         _ =>
                         {
                             if (_ == true)