+++ src/Narvalo.Core/Fx/Extensions/Maybe.g.cs
@@ -365,7 +365,7 @@ namespace Narvalo.Fx.Extensions.Internal
 
                 if (m != null)
                 {
-                    m.Apply(
+                    m.Invoke(
                         _ =>
                         {
                             if (_ == true)